requirements.txt
# LLM Provider SDKs
anthropic>=0.39.0
openai>=1.54.0
google-generativeai>=0.8.0

# Environment and Configuration
python-dotenv>=1.0.0

# XML Processing
lxml>=5.1.0

# Data Analysis and Metrics
numpy>=1.26.0
pandas>=2.2.0

# Text Similarity and NLP
sentence-transformers>=2.2.0
scikit-learn>=1.4.0

# HTTP and Async
httpx>=0.26.0
aiohttp>=3.9.0

# Type Checking and Validation
pydantic>=2.5.0
typing-extensions>=4.9.0

# Testing and Quality
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Logging and Progress
rich>=13.7.0
tqdm>=4.66.0

# JSON Schema Validation
jsonschema>=4.21.0

# Date/Time utilities
python-dateutil>=2.8.2
